# docs/swagger/swagger.yaml
basePath: /api/v0
definitions:
...
  pb.Cafe:
    properties:
      address:
        type: string
      api:
        type: string
      node:
        type: string
      peer:
        type: string
      protocol:
        type: string
      swarm:
        items:
          type: string
        type: array
      url:
        type: string
    type: object
  pb.CafeSession:
    properties:
      access:
        type: string
      cafe:
        $ref: '#/definitions/pb.Cafe'
        type: object
      exp:
        type: string
      id:
        type: string
      refresh:
        type: string
      rexp:
        type: string
      subject:
        type: string
      type:
        type: string
    type: object
  pb.CafeSessionList:
    properties:
      items:
        items:
          $ref: '#/definitions/pb.CafeSession'
        type: array
    type: object
...
  pb.Contact:
    properties:
      address:
        type: string
      avatar:
        type: string
      created:
        type: string
      id:
        type: string
      inboxes:
        items:
          $ref: '#/definitions/pb.Cafe'
        type: array
      threads:
        description: view info
        items:
          type: string
        type: array
      updated:
        type: string
      username:
        type: string
    type: object
  pb.ContactList:
    properties:
      items:
        items:
          $ref: '#/definitions/pb.Contact'
        type: array
    type: object
...
  pb.QueryResult:
    properties:
      date:
        type: string
      id:
        type: string
      local:
        type: boolean
      value:
        type: string
    type: object
...
host: 127.0.0.1:40600
info:
  contact:
    email: contact@textile.io
    name: Textile
    url: https://textile.io/
  description: Textile Node Local REST API Documentation
  license:
    name: MIT License
    url: https://github.com/textileio/go-textile/blob/master/LICENSE
  termsOfService: https://github.com/textileio/go-textile/blob/master/TERMS
  title: Textile Node API
  version: "0"
paths:
...
  /cafes:
    get:
      description: |-
        List info about all active cafe sessions. Cafes are other peers on the network
        who offer pinning, backup, and inbox services
      produces:
      - application/json
      responses:
        "200":
          description: cafe sessions
          schema:
            $ref: '#/definitions/pb.CafeSessionList'
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List info about all active cafe sessions
      tags:
      - cafes
    post:
      description: |-
        Registers with a cafe and saves an expiring service session token. An access
        token is required to register, and should be obtained separately from the target
        Cafe
      parameters:
      - description: cafe host
        in: header
        name: X-Textile-Args
        required: true
        type: string
      - default: token=
        description: 'token: An access token supplied by the Cafe'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: cafe session
          schema:
            $ref: '#/definitions/pb.CafeSession'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register with a Cafe
      tags:
      - cafes
...
  /contacts:
    get:
      description: Lists all, or thread-based, contacts.
      parameters:
      - default: thread=
        description: 'thread: Thread ID (omit for all known contacts)'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: contacts
          schema:
            $ref: '#/definitions/pb.ContactList'
            type: object
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List known contacts
      tags:
      - contacts
  /contacts/{id}:
    delete:
      description: Removes a known contact
      parameters:
      - description: contact id
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: ok
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Remove a contact
      tags:
      - contacts
    get:
      description: Gets information about a known contact
      parameters:
      - description: contact id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: contact
          schema:
            $ref: '#/definitions/pb.Contact'
            type: object
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get a known contact
      tags:
      - contacts
    put:
      consumes:
      - application/json
      description: |-
        Adds a contact to list of known local contacts. A common workflow is to pipe
        /contact/search results into this endpoint, just be sure you know what the results
        of the search are before adding!
      parameters:
      - description: contact id
        in: path
        name: id
        required: true
        type: string
      - description: contact
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/pb.Contact'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Add to known contacts
      tags:
      - contacts
  /contacts/search:
    post:
      description: Search for contacts known locally and on the network
      parameters:
      - default: local="false",limit=5,wait=5,peer=,address=,username=,events="false"
        description: 'local: Whether to only search local contacts, limit: Stops searching
          after limit results are found, wait: Stops searching after ''wait'' seconds
          have elapsed (max 10s), username: search by username string, peer: search
          by peer id string, address: search by account address string, events: Whether
          to emit Server-Sent Events (SSEvent) or plain JSON'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: results stream
          schema:
            $ref: '#/definitions/pb.QueryResult'
            type: object
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Search for contacts
      tags:
      - contacts
...
swagger: "2.0"